+++ library/src/main/java/com/prolificinteractive/materialcalendarview/CalendarDay.java
@@ -149,7 +149,6 @@ public final class CalendarDay implements Parcelable {
     void copyToMonthOnly(@NonNull Calendar calendar) {
         calendar.clear();
         calendar.set(year, month, 1);
-        calendar.getTimeInMillis();
     }
 
     /**
@@ -160,7 +159,6 @@ public final class CalendarDay implements Parcelable {
     public void copyTo(@NonNull Calendar calendar) {
         calendar.clear();
         calendar.set(year, month, day);
-        calendar.getTimeInMillis();
     }
 
     /**