+++ library/src/main/java/com/prolificinteractive/materialcalendarview/MaterialCalendarView.java
@@ -154,10 +154,14 @@ public class MaterialCalendarView extends ViewGroup {
     private final TextView title;
     private final DirectionButton buttonPast;
     private final DirectionButton buttonFuture;
-    private final ViewPager pager;
+    private final MonthPager pager;
     private final MonthPagerAdapter adapter;
     private CalendarDay currentMonth;
     private LinearLayout topbar;
+    /**
+     * Used for the dynamic calendar height.
+     */
+    private boolean mDynamicHeightEnabled;
 
     private final ArrayList<DayViewDecorator> dayViewDecorators = new ArrayList<>();
 
@@ -225,7 +229,7 @@ public class MaterialCalendarView extends ViewGroup {
         buttonPast = new DirectionButton(getContext());
         title = new TextView(getContext());
         buttonFuture = new DirectionButton(getContext());
-        pager = new ViewPager(getContext());
+        pager = new MonthPager(getContext());
 
         setupChildren();
 
@@ -315,7 +319,8 @@ public class MaterialCalendarView extends ViewGroup {
                     -1
             );
             if (firstDayOfWeek < 0) {
-                firstDayOfWeek = CalendarUtils.getInstance().getFirstDayOfWeek();
+                //Allowing use of Calendar.getInstance() here as a performance optimization
+                firstDayOfWeek = Calendar.getInstance().getFirstDayOfWeek();
             }
             setFirstDayOfWeek(firstDayOfWeek);
         } catch (Exception e) {
@@ -915,6 +920,7 @@ public class MaterialCalendarView extends ViewGroup {
         setTileSize(ss.tileSizePx);
         setTopbarVisible(ss.topbarVisible);
         setSelectionMode(ss.selectionMode);
+        setDynamicHeightEnabled(ss.dynamicHeightEnabled);
     }
 
     @Override
@@ -948,6 +954,7 @@ public class MaterialCalendarView extends ViewGroup {
         int tileSizePx = -1;
         boolean topbarVisible = true;
         int selectionMode = SELECTION_MODE_SINGLE;
+        boolean dynamicHeightEnabled = false;
 
         SavedState(Parcelable superState) {
             super(superState);
@@ -962,11 +969,12 @@ public class MaterialCalendarView extends ViewGroup {
             out.writeInt(showOtherDates);
             out.writeParcelable(minDate, 0);
             out.writeParcelable(maxDate, 0);
-            out.writeParcelableArray(selectedDates.toArray(new CalendarDay[selectedDates.size()]), 0);
+            out.writeTypedList(selectedDates);
             out.writeInt(firstDayOfWeek);
             out.writeInt(tileSizePx);
             out.writeInt(topbarVisible ? 1 : 0);
             out.writeInt(selectionMode);
+            out.writeInt(dynamicHeightEnabled ? 1 : 0);
         }
 
         public static final Parcelable.Creator<SavedState> CREATOR
@@ -989,12 +997,12 @@ public class MaterialCalendarView extends ViewGroup {
             ClassLoader loader = CalendarDay.class.getClassLoader();
             minDate = in.readParcelable(loader);
             maxDate = in.readParcelable(loader);
-            CalendarDay[] days = (CalendarDay[]) in.readParcelableArray(loader);
-            selectedDates.addAll(Arrays.asList(days));
+            in.readTypedList(selectedDates, CalendarDay.CREATOR);
             firstDayOfWeek = in.readInt();
             tileSizePx = in.readInt();
             topbarVisible = in.readInt() == 1;
             selectionMode = in.readInt();
+            dynamicHeightEnabled = in.readInt() == 1;
         }
     }
 
@@ -1016,6 +1024,27 @@ public class MaterialCalendarView extends ViewGroup {
     public int getFirstDayOfWeek() {
         return adapter.getFirstDayOfWeek();
     }
+
+    /**
+     * By default, the calendar will take up all the space needed to show any month (6 rows).
+     * By enabling dynamic height, the view will change height dependant on the visible month.
+     *
+     * This means months that only need 5 or 4 rows to show the entire month will only take up
+     * that many rows, and will grow and shrink as necessary.
+     *
+     * @param useDynamicHeight true to have the view different heights based on the visible month
+     */
+    public void setDynamicHeightEnabled(boolean useDynamicHeight) {
+        this.mDynamicHeightEnabled = useDynamicHeight;
+    }
+
+    /**
+     * @return the dynamic height state - true if enabled.
+     */
+    public boolean isDynamicHeightEnabled() {
+        return mDynamicHeightEnabled;
+    }
+
 
     /**
      * Add a collection of day decorators
@@ -1212,13 +1241,25 @@ public class MaterialCalendarView extends ViewGroup {
         final int desiredWidth = specWidthSize - getPaddingLeft() - getPaddingRight();
         final int desiredHeight = specHeightSize - getPaddingTop() - getPaddingBottom();
 
-        final int viewTileHieght = getTopbarVisible() ?
-                (MonthView.DEFAULT_MONTH_TILE_HEIGHT + 1) :
-                MonthView.DEFAULT_MONTH_TILE_HEIGHT;
+        int weekCount = MonthView.DEFAULT_MONTH_TILE_HEIGHT;
+
+        /*
+         * The default height of the calendar component is MonthView.DEFAULT_MONTH_TILE_HEIGHT rows,
+         * but we need to hide 1 if the actual date is 5 week long
+         */
+        if (mDynamicHeightEnabled && adapter != null && pager != null) {
+            Calendar cal = (Calendar) adapter.getItem(pager.getCurrentItem()).getCalendar().clone();
+            cal.set(Calendar.DAY_OF_MONTH, cal.getActualMaximum(Calendar.DAY_OF_MONTH));
+            //noinspection ResourceType
+            cal.setFirstDayOfWeek(getFirstDayOfWeek());
+            weekCount = cal.get(Calendar.WEEK_OF_MONTH) + 1; // + 1 because of the week days
+        }
+
+        final int viewTileHeight = getTopbarVisible() ? (weekCount + 1) : weekCount;
 
         //Calculate independent tile sizes for later
         int desiredTileWidth = desiredWidth / MonthView.DEFAULT_DAYS_IN_WEEK;
-        int desiredTileHeight = desiredHeight / viewTileHieght;
+        int desiredTileHeight = desiredHeight / viewTileHeight;
 
         int measureTileSize = -1;
 
@@ -1245,7 +1286,7 @@ public class MaterialCalendarView extends ViewGroup {
 
         //Calculate our size based off our measured tile size
         int measuredWidth = measureTileSize * MonthView.DEFAULT_DAYS_IN_WEEK;
-        int measuredHeight = measureTileSize * viewTileHieght;
+        int measuredHeight = measureTileSize * viewTileHeight;
 
         //Put padding back in from when we took it away
         measuredWidth += getPaddingLeft() + getPaddingRight();
@@ -1386,5 +1427,22 @@ public class MaterialCalendarView extends ViewGroup {
             super(MATCH_PARENT, tileHeight);
         }
 
+    }
+
+    /**
+     * Enable or disable the ability to swipe between months.
+     *
+     * @param pagingEnabled pass false to disable paging, true to enable (default)
+     */
+    public void setPagingEnabled(boolean pagingEnabled) {
+        pager.setPagingEnabled(pagingEnabled);
+        updateUi();
+    }
+
+    /**
+     * @return true if swiping months is enabled, false if disabled. Default is true.
+     */
+    public boolean isPagingEnabled() {
+        return pager.isPagingEnabled();
     }
 }